user can comment, delete comment, see posts and see comments

+++ MIntFFM/MIntFFM/src/main/java/com/beshoy/MIntFFM/user/UserRepo.java
@@ -7,10 +7,12 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface UserRepo extends JpaRepository<UserEntity, Integer> {
+public interface UserRepo extends JpaRepository<UserEntity, Long> {
     Optional<UserEntity> findByEmail(@NotBlank(message = "Email is required")
                                      @Email(message = "Invalid email format")
                                      String email);
 
     boolean existsByEmail(String email);
+
+
 }
